# notes-server-go/internal/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "拾记 API",
    "version": "0.4.0-go",
    "description": "Go + Gin implementation; existing REST v1 contract. Cookie/CSRF or scoped bearer authentication."
  },
  "paths": {
...
    "/api/v1/exports": {
      "get": {
        "tags": [
          "exports"
        ],
        "summary": "Listing",
        "operationId": "listing_api_v1_exports_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "exports"
        ],
        "summary": "Create",
        "operationId": "create_api_v1_exports_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ExportOptions"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/exports/{bundle_id}/download": {
      "get": {
        "tags": [
          "exports"
        ],
        "summary": "Download",
        "operationId": "download_api_v1_exports__bundle_id__download_get",
        "parameters": [
          {
            "name": "bundle_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Bundle Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "ExportOptions": {
        "properties": {
          "include_trash": {
            "type": "boolean",
            "title": "Include Trash",
            "default": true
          },
          "include_history": {
            "type": "boolean",
            "title": "Include History",
            "default": true
          }
        },
        "additionalProperties": false,
        "type": "object",
        "title": "ExportOptions"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}